# docs/openapi/delivery-plus-public.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Delivery Plus Public API",
    "version": "1.0.0",
    "description": "Public Gateway contract for the Delivery Plus food delivery platform. This specification exposes the canonical public paths behind the API Gateway and intentionally excludes internal-only service routes."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new account",
        "tags": [
          "auth"
        ],
        "operationId": "auth_post_api_auth_register",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login and receive a JWT",
        "tags": [
          "auth"
        ],
        "operationId": "auth_post_api_auth_login",
        "responses": {
          "200": {
            "description": "Authenticated response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "summary": "Verify an email address with a token",
        "tags": [
          "auth"
        ],
        "operationId": "auth_post_api_auth_verify_email",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": []
      }
    },
    "/api/auth/resend-verification": {
      "post": {
        "summary": "Resend email verification",
        "tags": [
          "auth"
        ],
        "operationId": "auth_post_api_auth_resend_verification",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": []
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Get the current authenticated user",
        "tags": [
          "auth"
        ],
        "operationId": "auth_get_api_auth_me",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Bad Request"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "password",
          "name"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AuthTokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}